+++ wildboar/__init__.py
@@ -15,4 +15,4 @@
 #
 # Authors: Isak Samsten
 
-__version__ = "1.0.6"
+__version__ = "1.0.6dev"